Bug Fix!
Before:
On the first video fragment, UVT calculation would add from 0ms to the end of the video

After:
On the first video fragment, UVT calculation adds the length of the fragment

Reasoning:
On first coding, assumed all videos would start at 0ms. If inputting only fragments starting after 0ms, would get numbers larger than intended.

+++ src/uvt_tracking_app/VideoViewFragment.java
@@ -4,7 +4,7 @@ public class VideoViewFragment implements Comparable<VideoViewFragment>{
 	
 	int viewStartTime;
 	int viewEndTime;
-	float totalViewTime;
+	int totalViewTime;
 	
 	VideoViewFragment(int startTime, int endTime){
 		if(endTime < startTime) {
